Initial work on moving to language provider
