+++ cli/.vscode/launch.json
@@ -24,7 +24,7 @@
 			"cwd": "${workspaceFolder:cli}",
 			"program": "${workspaceFolder:cli}/dist/index.js",
 			"sourceMaps": true,
-			"args": ["-d", "/Users/barry/Repos/ibmi-company_system", "--verbose", "--mcp", "5500"],
+			"args": ["-d", "/Users/barry/Repos/ibmi-company_system", "--verbose", "-bf", "make"],
 			"preLaunchTask": {
 				"type": "npm",
 				"script": "webpack:dev"
