+++ cli/src/utils.ts
@@ -315,3 +315,13 @@ export function globalEntryIsValid(fullPath: string, search: string, ignoreBase?
 
 	return false;
 }
+
+export function trimQuotes(input: string|boolean, value = `'`) {
+	if (typeof input === `string`) {
+		if (input[0] === value) input = input.substring(1);
+		if (input[input.length - 1] === value) input = input.substring(0, input.length - 1);
+		return input;
+	} else {
+		return '';
+	}
+}
